agent/requirements.txt
livekit-agents[openai,silero,deepgram,elevenlabs]~=1.0
python-dotenv==1.0.1
httpx==0.27.2